+++ .env
@@ -1,0 +1,2 @@
+TODO_PORT="1818"
+TODO_DBFILE="./scheduler.db"
