+++ internal/models/models.go
@@ -1,1 +1,10 @@
 package models
+
+// Task представляет структуру задачи
+type Task struct {
+	ID      string `json:"id"`
+	Date    string `json:"date"`
+	Title   string `json:"title"`
+	Comment string `json:"comment"`
+	Repeat  string `json:"repeat"`
+}
